バリデーションを切り出し
