+++ functions/src/handlers/get-task-handler.ts
@@ -1,5 +1,4 @@
 import { APIGatewayEvent } from 'aws-lambda';
-import { z } from 'zod';
 import { httpResponse, LambdaResponse } from './http/http-response';
 
 import { getTaskUseCase } from '../usecases/get-task-usecase';
@@ -8,25 +7,14 @@ import {
   RequestHandlerWithoutContext,
 } from './factory/handler-factory';
 import { HttpStatus } from './http/http-status';
-import { ErrorCode } from '../common/errors/error-codes';
-import { AppError } from '../common/errors/app-errors';
-
-export const EventSchema = z.object({
-  pathParameters: z.object({ id: z.string().uuid() }),
-});
+import { validateEventWithPathParams } from './http/validators';
 
 const requestHandler: RequestHandlerWithoutContext = async (
   event: APIGatewayEvent,
 ): Promise<LambdaResponse> => {
-  const eventResult = EventSchema.safeParse(event);
+  const pathParams = validateEventWithPathParams(event).pathParameters;
 
-  if (!eventResult.success) {
-    throw new AppError(ErrorCode.INVALID_REQUEST);
-  }
-
-  const { pathParameters } = eventResult.data;
-
-  const task = await getTaskUseCase(pathParameters.id);
+  const task = await getTaskUseCase(pathParams.id);
   return httpResponse(HttpStatus.OK).withBody(task);
 };
